refactor(vestad): extract state.rs and settings.rs from serve.rs so handler modules import a module below them
